requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

aiortc==1.9.0
async-timeout==5.0.1
bcrypt==4.2.1
channels==4.2.0
Django==5.1.4
djangorestframework==3.15.2
libtorrent==2.0.9
pillow==11.1.0
python-dotenv==1.0.1
requests==2.32.3
rich==9.13.0
torrent-cli==0.0.9
torrentstream==0.2.2
websockets==14.1
yagmail==0.15.293